LayoutTests/imported/w3c/web-platform-tests/html/infrastructure/safe-passing-of-structured-data/structured_clone_bigint-expected.txt

CONSOLE MESSAGE: line 31: SyntaxError: No identifiers allowed directly after numeric literal

FAIL 2.7 Safe passing of structured data SyntaxError: No identifiers allowed directly after numeric literal
